% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugNet.R
\name{AdjMatrix_drugs}
\alias{AdjMatrix_drugs}
\title{Generate drug adjacency matrix}
\usage{
AdjMatrix_drugs(
  x,
  weighted = c("TRUE", "T", "FALSE", "F"),
  weight.type = c("ATCsim", "STsim", "SEsim", "integrated_score", "integrated_score2")
)
}
\arguments{
\item{x}{drug-drug similarity feature file generated from
\code{DrugNetFeature}}

\item{weighted}{\code{TRUE/FALSE}, choose if the edge of drug network should
be weighted by feature values. The abbreviationS "\code{T/F}" are also
acceptable}

\item{weight.type}{if parameter "\code{weighted = TRUE/T}", choose the
feature type to weight the edge of drug-drug network. The options are:
"ATCsim","STsim","SEsim","integrated_score","integrated_score2"}
}
\value{
\code{AdjMatrix_drugs} Drug adjacency matrix generated from drug-drug
  association network
}
\description{
The function \code{AdjMatrix_drugs} is to generate adjacency
  matrix for drug-drug interaction network weighted by the integrated
  pharmacological similarity values obtained from \code{DrugNetFeature}.
}
\examples{

# drugfeature = read.table(paste0(load_dir, "drug_net/features.csv", sep =
# ",", header = TRUE, stringsAsFactors = FALSE))
\dontrun{
AdjMatrix_drugs(x = drugfeature, weighted = TRUE, weight.type = "integrated_score")
}
}
