% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugNetFeatures.R
\name{ATC.sim}
\alias{ATC.sim}
\title{Drug ATC Code-Based Similarity Function}
\usage{
ATC.sim(druglist = NULL, load_dir)
}
\arguments{
\item{druglist}{a list of drugs in \code{data frame} format}

\item{load_dir}{path to load or save modeling data files}
}
\value{
\code{ATC.sim} returns \code{ATCsim.matrix}, drug-drug ATC code based
  similarity matrix. Row name and column name are drug names, values
  represent drug-drug ATC code based similarity. Matrix will be saved in
  provided path \code{load_dir} under \emph{"/drug_net/atc_sim/"} folder, under the
  name of "ATCsim.txt"
}
\description{
The function \code{ATC.sim} is used to calculate drug-drug ATC code-based
  similarity.
}
\details{
\code{ATC.sim} integrated \code{\link{ATC.Dis}} and
  \code{\link{drugATCprep}} functions. \code{druglist} in \code{data.frame}
  should be provided. Their ATC code(s) will be found via \code{drugATCprep}.

\code{ATC.sim} will compute shortest distance between drugs based on
  the shortest distance between ATC codes  within ATC hierarchical network:
  \deqn{sim_{ATC}(t_{i},t_{j}) = w(t_{i})*w(t_{j})e^(-\gamma*d(t_{i},t_{j}))}
  where t_{i} and t_{j} denote the ATC code i and j, d(t_{i},t_{j}) denotes the
  shortest distance between t_{i} and t_{j} and computes with function
  \code{\link{ATC.Dis}}; w(t_{i}) and w(t_{j}) represent the weights of the
  corresponding ATC codes, and are defined as the inverse of ATC-code
  frequencies. \eqn{\gamma} is a pre-defined parameter (set as 0.25). For
  drugs with multiple ATC codes, average ATC code similarities will be
  considered.

ATC code based similarity between two drugs will be calculated and
  saved in \code{load_dir} path provided by user.
}
\examples{

\dontrun{
druglist = data.frame(drug = "Sorafenib")
load_dir = "G:/lab/DCcomboNet/Rpackage/input_data/"
ATC.sim(druglist = druglist, load_dir = load_dir)
}

}
\seealso{
\code{\link{drugATCprep}} to prepare input drug_atc file,
  \code{\link{ATC.Dis}} for computing intermediated ATC distance matrix
}
