% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneNet.R
\name{cancerGene}
\alias{cancerGene}
\title{Generate cancer gene list}
\usage{
cancerGene(cellline, load_dir)
}
\arguments{
\item{cellline}{cancer cell lines name, if requested cell line name is
included in the default cell line list, cancer sample specific expressed
gene list will be computed and saved in given path; otherwise
\emph{ERROR} report and \emph{SUGGESTION} will be returned}

\item{load_dir}{defined the path to load and save model data files}
}
\value{
\code{cell_genelist} cancer cell line specific expressed gene list,
  which will be returned and saved in provided \code{load_dir} under the
  folder named "\emph{gene_network}" in \emph{.txt} format, file will be
  named as " "\code{cellline}"_genelist.txt" name you provide
}
\description{
The function \code{cancerGene} is to generate cancer sample
  specific expressed gene list from CCLE database. Cancer sample specific
  expressed genes are obtained by comparing gene expression of certain cancer
  sample with the rest, fold-change above 1.5 times is considered as specific
  expressed genes.
}
\examples{
\dontrun{
cellline = "HEPG2"
load_dir = "G:/lab/DCcomboNet/Rpackage/input_data/"

cancerGene(cellline = cellline, load_dir = load_dir)
}
}
