% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{cellline_lib}
\alias{cellline_lib}
\title{Cancer cell lines in CCLE database}
\description{
Avaliable cancer cell lines in CCLE database, checking if user inputs cell
line name in including in CCLE so Level 2 model can be called without
requiring extra cancer specific expressed gene table
}
