% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{celllineDEG_lib}
\alias{celllineDEG_lib}
\title{Cancer cell lines in LINCS database}
\description{
Avaliable cancer cell lines in LINCS database, checking if user inputs cell
line name in including in LINCS so Level 2 model can be called without
requiring extra drug induced differentially expressed gene table
}
