% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prediction.R
\name{DComboNet}
\alias{DComboNet}
\title{Prediction drug combinations}
\usage{
DComboNet(
  load_dir,
  resultdir,
  model = c("L1", "L2"),
  drugcandidate = NULL,
  manual_input = c("TRUE", "T", "FALSE", "F"),
  CDK_FP = NULL,
  pubchemFP = NULL,
  MACCS_FP = NULL,
  drugnetWeight = c("TRUE", "T", "FALSE", "F"),
  featuretype = c("ATCsim", "STsim", "SEsim", "integrated_score"),
  drugtarget = NULL,
  druggene = NULL,
  dataset = NULL,
  cellline = NULL,
  treatment_time = NULL,
  foldchange_DEG = NULL,
  pvalue_DEG = NULL,
  foldchange_DEP = NULL,
  pvalue_DEP = NULL,
  drugDEG = NULL,
  drugDEP = NULL,
  cancergene = NULL,
  dtweight = 2,
  dgweight = 1,
  dDEGweight = 1,
  dgpweight = 1,
  dDEpweight = 1,
  gpweight = 1,
  x = 0.5,
  y = 0.5,
  z = 0,
  A = 0.5,
  B = 0.5,
  eta = 1,
  r = 0.7
)
}
\arguments{
\item{load_dir}{path to load or save modelin data files}

\item{resultdir}{path to save result files}

\item{model}{choose model type, two models here to choose: "L1", "L2"}

\item{drugcandidate}{drug and its corresponding "drugbank ID" in
\code{data.frame}}

\item{manual_input}{logical (\code{TRUE(T)/FALSE(F)}), if user wants to
manually input drug(s)}

\item{CDK_FP}{Drug fingerprint matrix generated from PEDAL}

\item{pubchemFP}{Drug Pubchem fingerprint matrix generated from PEDAL}

\item{MACCS_FP}{Drug MACCS fingerprint matrix generated from PEDAL}

\item{drugnetWeight}{logical (\code{TRUE(T)/FALSE(F)}), choose if the edge of
drug network should be weighted by feature values. The abbreviationS
"\code{T/F}" are also acceptable}

\item{featuretype}{if parameter "\code{weighted = TRUE/T}", choose the
feature type to weight the edge of drug-drug network. The options are:
"ATCsim","STsim","SEsim","integrated_score","integrated_score2"}

\item{drugtarget}{drug-target gene interaction file in \code{data.frame}
format (otherwise \code{as.data.frame} can be used to transform the format
of file). Target gene names should convert to official gene name (i.e. HGNC
gene symbol)}

\item{druggene}{drug-gene interaction file (Optional). drug and genes that
either induced by drug or may influence drug effect can be provided for
construct gene network for level one model. Genes can be collected from
publications or from IPA tool. Gene name should be converted to official
gene names (i.e. HGNC gene symbols)}

\item{dataset}{GEO ID of LINCS datasets, two datasets are provided here:
"70138" and "92742"}

\item{cellline}{cancer cell lines name. If level two model is called
(\code{model = "L2"} in \code{\link{DComboNet}} or
\code{\link{L2.LOCOCV}}). If requested cell line is included in the default
cancer cell line list, \code{dDEG} do not need to provided unless
particular drug induced differentially expressed gene list you want to
include in the gene network, then provided \code{dDEG} in
\code{as.data.frame} format. If the cancer cell line is not in the current
cell line list, you should provide \code{cellline} and \code{dDEG} for
construct specific gene network, otherwise \emph{ERROR} information will be
returned}

\item{treatment_time}{drug treatment time provided in LINCS database. For
example, MCF7 cell line (\code{cellline = "MCF7"}) provided two time points
in LINCS GSE\code{70138} dataset - 6 hour and 12 hour - \code{6} or
\code{12} should be taken as input here, as numeric or character}

\item{foldchange_DEG}{numeric, fold change of gene expression between
drug-treated cancer cell line and DMSO treated cell line}

\item{pvalue_DEG}{numeric, p-values corresponding to the t-statistics}

\item{foldchange_DEP}{numeric, fold change of enriched pathways via GSVA
between drug-treated cancer cell line and DMSO treated cell line}

\item{pvalue_DEP}{numeric, p-values corresponding to the t-statistics}

\item{drugDEG}{if \code{cellline} is not included in provided cancer cell
line list, in \code{as.data.frame} format must be provided, otherwise
\emph{ERROR} information will be returned}

\item{drugDEP}{if \code{cellline} is not included in provided cancer cell
line list, in \code{as.data.frame} format must be provided, otherwise
\emph{ERROR} information will be returned}

\item{cancergene}{cancer_gene a list of cancer related genes. If Level one
model is called (\code{model = "L1"} in \code{\link{DComboNet}} or
\code{\link{L1.LOCOCV}}), \code{cancer_gene} list is provided within the
function, it can be \code{NULL}. In level two model, if cell line name is
included in our provided cancer specific gene lists, \code{NULL} can take
as input, otherwise \code{cancer_gene} should be prepared in
\code{data.frame} format}

\item{dtweight}{the weight of drug-target gene edges}

\item{dgweight}{the weight of drug-related gene edge}

\item{dDEGweight}{the weight of drug-differentially expressed genes(DEGs)
edge}

\item{dgpweight}{numeric, weight of drug-pathway associations through drug
related genes}

\item{dDEpweight}{numeric, weight of drug-drug-differentially expressed
pathway edges}

\item{gpweight}{the weight of gene-pathway edges}

\item{x}{numeric, the probability that the inter-network crossing event
happens between drug and gene nodes}

\item{y}{numeric, the probability that the inter-network crossing event
happens between drug and pathway nodes}

\item{z}{numeric, the probability that the inter-network crossing event
happens between gene and pathway nodes}

\item{A}{numeric, the probability of jumping within drugnet, when the
inter-network crossing events between drug and gene nodes and between drug
and pathway nodes are neither existing}

\item{B}{numeric, the probability of jumping within genenet, when the
inter-network crossing events between drug and gene nodes and between gene
and pathway nodes are neither existing}

\item{eta}{numeric parameter to controls the probability of restarting in the
corresponding network}

\item{r}{numeric, global restart parameter}
}
\value{
Ranked result tables. Each drug seed will create three files that
  contains global proximity between drug seed and other drug/gene/pathway
  nodes computed from drug-gene/pathway interaction network. File will be
  saved under the path \code{resultdir} provided by user.
}
\description{
the function \code{DComboNet} is to predict possible drug to
  pair with the drug candidate without expression data
}
\seealso{
\code{\link{L1.LOCOCV}}, \code{\link{L2.LOCOCV}}
}
