% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_DEpathway_preparation.R
\name{DEG_DEP_preparation}
\alias{DEG_DEP_preparation}
\title{Drug induced differentially expressed gene/pathway list preparation}
\usage{
DEG_DEP_preparation(
  druglist,
  dataset,
  cellline,
  treatment_time,
  core,
  load_dir
)
}
\arguments{
\item{druglist}{drug and its corresponding "drugbank ID" in
\code{data.frame}}

\item{dataset}{GEO ID of LINCS datasets, two datasets are provided here:
"70138" and "92742"}

\item{cellline}{cancer cell lines name. If level two model is called
(\code{model = "L2"} in \code{DComboNet} or \code{L2.LOCOCV}). If requested
cell line is included in the default cancer cell line list, \code{dDEG} do
not need to provided unless particular drug induced differentially
expressed gene list you want to include in the gene network, then provided
\code{dDEG} in \code{as.data.frame} format. If the cancer cell line is not
in the current cell line list, you should provide \code{cellline} and
\code{dDEG} for construct specific gene network, otherwise \emph{ERROR}
information will be returned}

\item{treatment_time}{drug treatment time provided in LINCS database. For
example, MCF7 cell line (\code{cellline = "MCF7"}) provided two time points
in LINCS GSE70138 dataset - 6 hour and 12 hour - \code{6} or \code{12}
should be taken as input here}

\item{core}{numeric, the number of core used for parallel running \code{gsva}}

\item{load_dir}{path to load or save modeling data files}
}
\description{
The function \code{DEG_DEP_preparation} is to prepare drug
  induced differentially expressed gene table and pathway table for level two
  model(\code{model = "L2"}) from LINCS database. Differentially expressed
  genes were selected by functions \code{lmFIt} and \code{eBayes} in the
  \code{Limma} package. Differential regulated pathways were obtained by
  \code{gsva} function from \code{GSVA} package.
}
\examples{
druglist = data.frame(Druglist = c("Sorafenib", "Sunitinib"))
dataset = "92742" # which dataset to use
cellline = "HEPG2"
t = 6
\dontrun{
# access the number of available cores to run GSVA
library(parallel)
num_cores<-detectCores(logical=F)
}
num_cores = 2
load_dir = "G:/lab/DCcomboNet/Rpackage/input_data/"
\dontrun{
DEG_DEP_preparation(druglist = druglist, dataset = dataset, cellline =
cellline, treatment_time = t, core = num_cores, load_dir = load_dir)
}

}
