% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{dg_lib}
\alias{dg_lib}
\title{Drugs and their related genes}
\description{
Drugs in pre-constructed network and their related genes extracted from IPA
(Ingenuity® Pathway Analysis) tool.
}
