% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugNetFeatures.R
\name{drugATCprep}
\alias{drugATCprep}
\title{Drug ATC Code Distance file preparation}
\usage{
drugATCprep(druglist = NULL, load_dir)
}
\arguments{
\item{druglist}{a list of drugs in \code{data.frame} format}

\item{load_dir}{path to load or save modelling data files}
}
\value{
\code{drug_atc} a data frame of drug and its corresponding ATC code
  in ATC system
}
\description{
The function \code{drugATCprep} is used for preparing input
  files for calculating drug ATC code distance in Anatomical Therapeutic
  Chemical Classification System (ATC system). A drug-ATC file is packed in
  the function for newly inputed drug(s) to find its ATC codes
}
\examples{
\dontrun{
druglist = data.frame(drug = "Sorafenib")
load_dir = "G:/lab/DCcomboNet/Rpackage/input_data/"
drugATCprep(druglist = druglist, load_dir = load_dir)}

}
