% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_DEpathway_preparation.R
\name{drugDEG_preparation}
\alias{drugDEG_preparation}
\title{Drug induced differentially expressed gene list
preparation}
\usage{
drugDEG_preparation(
  cellline,
  dataset,
  treatment_time,
  foldchange,
  pvalue,
  load_dir
)
}
\arguments{
\item{cellline}{cancer cell lines name. If level two model is called
(\code{model = "L2"} in \code{DComboNet} or \code{L2.LOCOCV}). If requested
cell line is included in the default cancer cell line list, \code{dDEG} do
not need to provided unless particular drug induced differentially
expressed gene list you want to include in the gene network, then provided
\code{dDEG} in \code{as.data.frame} format. If the cancer cell line is not
in the current cell line list, you should provide \code{cellline} and
\code{dDEG} for construct specific gene network, otherwise \emph{ERROR}
information will be returned}

\item{dataset}{GEO ID of LINCS datasets, two datasets are provided here:
"70138" and "92742"}

\item{treatment_time}{drug treatment time provided in LINCS database. For
example, MCF7 cell line (\code{cellline = "MCF7"}) provided two time points
in LINCS GSE70138 dataset - 6 hour and 12 hour - \code{6} or \code{12}
should be taken as input here}

\item{foldchange}{numeric, fold change of gene expression between
drug-treated cancer cell line and DMSO treated cell line}

\item{pvalue}{numeric, p-values corresponding to the t-statistics}

\item{load_dir}{path to load or save modeling data files}
}
\description{
The function \code{drugDEG_preparation} is to prepare drug
  induced differentially expressed gene table and pathway table for level two
  model(\code{model = "L2"}). If drug or cancer cell line you are interested
  in are not included in our dataset but contained in LINCS datasets
  ("GSE70138" or "GSE92742"), function \code{DEG_DEP_preparation} provides
  integrated analysis tools to generate matching files.
}
\examples{
dataset = "92742" # which dataset to use
cellline = "HEPG2"
t = 6
FC = 0.5
p = 0.05
load_dir = "G:/lab/DCcomboNet/Rpackage/input_data/"
\dontrun{
drugDEG_preparation(cellline = cellline, dataset = dataset, treatment_time =
t, foldchange = FC, pvalue = p, load_dir = load_dir)
}

}
