% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugGeneNet.R
\name{drugGeneNet.L1}
\alias{drugGeneNet.L1}
\title{Generate adjacency matrix from drug-gene interaction networks
for level one model}
\usage{
drugGeneNet.L1(
  dt = NULL,
  dg = NULL,
  drugAdj,
  geneNet,
  dtweight,
  dgweight,
  load_dir
)
}
\arguments{
\item{dt}{drug-target gene interaction file in \strong{data.frame} format
(otherwise \code{as.data.frame} can be used to transform the format of
file). Target gene names should convert to official gene name (i.e. HGNC
gene symbol).}

\item{dg}{drug-gene interaction file (Optional). drug and genes that either
induced by drug or may influence drug effect can be provided for construct
gene network for level one model. Genes can be collected from publications
or from IPA tool. Gene name should be converted to official gene names
(i.e. HGNC gene symbols).}

\item{drugAdj}{drug adjacency matrix generated from drug-drug
association network via \code{AdjMatrix_drugs}}

\item{geneNet}{gene-gene association network adjacency matrix (\bold{sparse}
matrix format)}

\item{dtweight}{the weight of drug-target gene edges}

\item{dgweight}{the weight of drug-related gene edge}

\item{load_dir}{path to load or save modeling data files}
}
\value{
\code{druggeneMatrix} drug-gene adjacency matrix
}
\description{
The function \code{drugGeneAdj} is to generate the adjacency
  matrix from gene-gene interaction network
}
\examples{
#
# loading files
\dontrun{
load_dir = "G:/lab/DCcomboNet/Rpackage/input_data/"
drugnet_feature = read.table(paste0(load_dir,"drug_net/features.csv"),
sep=",",header=TRUE, stringsAsFactors = FALSE)
drugnet_feature = drugnet_feature[drugnet_feature$integrated_score2 >=0.2,]
drugtarget = read.table(paste0(load_dir,"data/drugtarget.csv"), sep =
",",header = TRUE, stringsAsFactors = FALSE)
druggene = data.frame(drug = "Sorafenib", target  = "RAF1")
drugnet_adj = AdjMatrix_drugs(x = drugnet_feature, weighted = TRUE,
weight.type = "integrated_score")
gene.net = geneNet(dt = drugtarget, dg = druggene, load_dir = load_dir)
dtweight = 2
dgweight = 1
#
# construct drug-gene adjacency matrix for level one model
drugGeneNet.L1(dt=drugtarget, dg=druggene, drugAdj=drugnet_adj, geneNet =
gene.net, dtweight = dtweight, dgweight = dgweight, load_dir = load_dir)
}

}
