% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugGeneNet.R
\name{drugGeneNet.L2}
\alias{drugGeneNet.L2}
\title{Generate the bipartite matrix from drug-gene interaction
networks for level two model}
\usage{
drugGeneNet.L2(
  dt = NULL,
  dg = NULL,
  dDEG = NULL,
  drugAdj,
  geneNet,
  dataset = NULL,
  cellline,
  treatment_time = NULL,
  dtweight,
  dgweight,
  dDEGweight,
  load_dir
)
}
\arguments{
\item{dt}{drug-target gene interaction file in \strong{data.frame} format
(otherwise \code{as.data.frame} can be used to transform the format of
file). Target gene names should convert to official gene name (i.e. HGNC
gene symbol).}

\item{dg}{drug-gene interaction file (Optional). drug and genes that either
induced by drug or may influence drug effect can be provided for construct
gene network for level one model. Genes can be collected from publications
or from IPA tool. Gene name should be converted to official gene names
(i.e. HGNC gene symbols).}

\item{dDEG}{if \code{cellline} is not included in provided cancer cell line
list, in \code{as.data.frame} format must be provided, otherwise
\emph{ERROR} information will be returned}

\item{drugAdj}{drug adjacency matrix generated from drug-drug association
network via \code{AdjMatrix_drugs}}

\item{geneNet}{gene-gene association network adjacency matrix (\bold{sparse}
matrix format)}

\item{dataset}{GEO ID of LINCS datasets, two datasets are provided here:
"70138" and "92742"}

\item{cellline}{cancer cell lines name. If level two model is called
(\code{model = "L2"} in \code{DComboNet} or \code{L2.LOCOCV}). If requested
cell line is included in the default cancer cell line list, \code{dDEG} do
not need to provided unless particular drug induced differentially
expressed gene list you want to include in the gene network, then provided
\code{dDEG} in \code{as.data.frame} format. If the cancer cell line is not
in the current cell line list, you should provide \code{cellline} and
\code{dDEG} for construct specific gene network, otherwise \emph{ERROR}
information will be returned}

\item{treatment_time}{drug treatment time provided in LINCS database. For
example, MCF7 cell line (\code{cellline = "MCF7"}) provided two time points
in LINCS GSE70138 dataset - 6 hour and 12 hour - \code{6} or \code{12}
should be taken as input here, as numeric or character}

\item{dtweight}{the weight of drug-target gene edges}

\item{dgweight}{the weight of drug-related gene edge}

\item{dDEGweight}{the weight of drug-differentially expressed genes(DEGs)
edge}

\item{load_dir}{path to load or save modeling data files}
}
\value{
\code{druggeneMatrix} drug-gene adjacency matrix
}
\description{
The function \code{drugGeneNet.L2} is to generate the adjacency
  matrix from gene-gene interaction network for level two model
}
\examples{
#
# loading files

\dontrun{
  load_dir = "G:/lab/DCcomboNet/Rpackage/input_data/"
drugnet_feature = read.table(paste0(load_dir,"drug_net/features.csv"),
sep=",",header=TRUE, stringsAsFactors = FALSE)
drugnet_feature = drugnet_feature[drugnet_feature$integrated_score2 >=0.2,]
drugtarget = read.table(paste0(load_dir,"data/drugtarget.csv"), sep =
",",header = TRUE, stringsAsFactors = FALSE)
druggene = data.frame(drug = "Sorafenib", target  = "RAF1")
drugnet_adj = AdjMatrix_drugs(x = drugnet_feature, weighted = TRUE,
weight.type = "integrated_score")
drugDEG = NULL
# or drugDEG = data.frame(drug = "Sorafenib", target = c("FOSL1", "TUBB6"))
dataset = "92742"
cellline = "HEPG2"
t="6"
cancergene = NULL
# or cancergene = c("AKT3","CDH2")
gene.net = geneNet(dt = drugtarget, dg = druggene, dDEG = drugDEG, dataset=
dataset, cellline = cellline, treatment_time = t, cancer_gene = cancergene,
load_dir = load_dir)
dtweight = 2
dgweight = 1
dDEGweight = 1
#
# construct drug-gene adjacency matrix for level two model
drugGeneNet.L2(dt=drugtarget, dg=druggene, dDEG = drugDEG,
drugAdj=drugnet_adj, geneNet = gene.net, dataset = dataset, cellline =
cellline, treatment_time = t, dtweight = dtweight, dgweight = dgweight,
dDEGweight = dDEGweight, load_dir = load_dir)
}

}
