% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{drugpair_lib}
\alias{drugpair_lib}
\title{drugpair_lib}
\description{
all drug pairs in pre-constructed network, including known drug combinations
marked as \code{P} and randomly reshuffled pairs marked as \code{N} in "TAG"
column.
}
