% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugPathwayNet.R
\name{DrugPathwayMatrix.L1}
\alias{DrugPathwayMatrix.L1}
\title{Generate adjacency matrix from drug-pathway interaction networks for level
one model}
\usage{
DrugPathwayMatrix.L1(
  dt = NULL,
  dg = NULL,
  drugAdj,
  PathwayNetwork,
  dgpweight,
  load_dir
)
}
\arguments{
\item{dt}{drug-target gene interaction file in \strong{data.frame} format
(otherwise \code{as.data.frame} can be used to transform the format of
file). Target gene names should convert to official gene name (i.e. HGNC
gene symbol).}

\item{dg}{drug-gene interaction file (Optional). drug and genes that either
induced by drug or may influence drug effect can be provided for construct
gene network for level one model. Genes can be collected from publications
or from IPA tool. Gene name should be converted to official gene names
(i.e. HGNC gene symbols).}

\item{drugAdj}{drug adjacency matrix (sparse matrix) generated from drug-drug
association network via \code{AdjMatrix_drugs}}

\item{PathwayNetwork}{\code{data.frame} of pathway pathway association network
network, sparse matrix generated via \code{pathwayNet.Adj}}

\item{dgpweight}{numeric, weight of drug-pathway associations through drug
related genes}

\item{load_dir}{path to load or save modeling data files}
}
\value{
\code{drugpathwayMatrix} drug-pathway adjacency matrix (sparse
  matrix)
}
\description{
The function "DrugPathwayMatrix.L1" is to generate the adjacency
  matrix from drug-pathway interaction network for level one model
}
\examples{
#
# loading files
\dontrun{
load_dir = "G:/lab/DCcomboNet/Rpackage/input_data/"
drugnet_feature = read.table(paste0(load_dir,"drug_net/features.csv"),
sep=",",header=TRUE, stringsAsFactors = FALSE)
drugtarget = read.table(paste0(load_dir,"data/drugtarget.csv"), sep =
",",header = TRUE, stringsAsFactors = FALSE)
druggene = data.frame(drug = "Sorafenib", target  = "RAF1")
# or druggene = NULL
drugnet_adj = AdjMatrix_drugs(x = drugnet_feature, weighted = TRUE,
weight.type = "integrated_score")
PathwayNetwork = read.table(paste0(load_dir,"pathway/WWI.txt"), sep = "\t",
header =TRUE, stringsAsFactors = FALSE)
dgpweight = 1
#
# construct drug-pathway adjacency matrix for level one model#'
DrugPathwayMatrix.L1(dt = drugtarget, dg = druggene, drugAdj = drugnet_adj,
dgpweight = dgpweight, PathwayNetwork = PathwayNetwork, load_dir = load_dir)
}

}
