% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugPathwayNet.R
\name{DrugPathwayMatrix.L2}
\alias{DrugPathwayMatrix.L2}
\title{Generate the bipartite matrix from drug-pathway interaction networks}
\usage{
DrugPathwayMatrix.L2(
  dt = NULL,
  dg = NULL,
  dDEG = NULL,
  dataset = NULL,
  cellline,
  treatment_time = NULL,
  drugDEP = NULL,
  drugAdj,
  PathwayNetwork,
  dDEpweight = 2,
  load_dir
)
}
\arguments{
\item{dt}{drug-target gene interaction file in \strong{data.frame} format
(otherwise \code{as.data.frame} can be used to transform the format of
file). Target gene names should convert to official gene name (i.e. HGNC
gene symbol).}

\item{dDEG}{if \code{cellline} is not included in provided cancer cell line
list, in \code{as.data.frame} format must be provided, otherwise
\emph{ERROR} information will be returned}

\item{dataset}{GEO ID of LINCS datasets, two datasets are provided here:
"70138" and "92742"}

\item{cellline}{cancer cell lines name. If level two model is called
(\code{model = "L2"} in \code{DComboNet} or \code{L2.LOCOCV}). If requested
cell line is included in the default cancer cell line list, \code{dDEG} do
not need to provided unless particular drug induced differentially
expressed gene list you want to include in the gene network, then provided
\code{dDEG} in \code{as.data.frame} format. If the cancer cell line is not
in the current cell line list, you should provide \code{cellline} and
\code{dDEG} for construct specific gene network, otherwise \emph{ERROR}
information will be returned}

\item{treatment_time}{drug treatment time provided in LINCS database. For
example, MCF7 cell line (\code{cellline = "MCF7"}) provided two time points
in LINCS GSE70138 dataset - 6 hour and 12 hour - \code{6} or \code{12}
should be taken as input here, as numeric or character}

\item{drugDEP}{if \code{cellline} is not included in provided cancer cell
line list, in \code{as.data.frame} format must be provided, otherwise
\emph{ERROR} information will be returned}

\item{drugAdj}{drug adjacency matrix (sparse matrix) generated from drug-drug
association network via \code{AdjMatrix_drugs}}

\item{PathwayNetwork}{\code{data.frame} of pathway pathway association
network network, sparse matrix generated via \code{pathwayNet.Adj}}

\item{dDEpweight}{numeric, weight of drug-drug-differentially expressed
pathway edges}

\item{load_dir}{path to load or save modeling data files}
}
\value{
\code{drugpathwayMatrix} drug-pathway adjacency matrix (sparse
  matrix)
}
\description{
The function \code{DrugPathwayMatrix.L2} is to generate
  adjacency matrix from drug-pathway interaction network for level two model.
}
\examples{
#
# loading files
\dontrun{
load_dir = "G:/lab/DCcomboNet/Rpackage/input_data/"
drugnet_feature = read.table(paste0(load_dir,"drug_net/features.csv"),
sep=",",header=TRUE, stringsAsFactors = FALSE)
drugnet_feature = drugnet_feature[drugnet_feature$integrated_score2 >=0.2,]
drugtarget = read.table(paste0(load_dir,"data/drugtarget.csv"), sep =
",",header = TRUE, stringsAsFactors = FALSE)
drugDEG = NULL
# or drugDEG = data.frame(drug = "Sorafenib", target = c("FOSL1", "TUBB6"))
dataset = "92742"
cellline = "HEPG2"
t="6"
drugnet_adj = AdjMatrix_drugs(x = drugnet_feature, weighted = TRUE,
weight.type = "integrated_score")
PathwayNetwork = read.table(paste0(load_dir,"pathway/WWI.txt"), sep = "\t",
header =TRUE, stringsAsFactors = FALSE)
dDEpweight = 1
#
# construct drug-pathway adjacency matrix for level two model
DrugPathwayMatrix.L2(dt = drugtarget, dDEG = drugDEG, dataset=dataset,
cellline = cellline, treatment_time=t, drugAdj = drugnet_adj, dDEpweight =
dDEpweight, PathwayNetwork = PathwayNetwork, load_dir = load_dir)
}

}
