% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{drugselib}
\alias{drugselib}
\title{Drugs and their side effect terms}
\description{
all drugs in pre-constructed network and their corresponding side effect
terms extracted from SIDER(v4) database
}
