% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugNetFeatures.R
\name{drugSEprep}
\alias{drugSEprep}
\title{Drug side effect similarity file preparation}
\usage{
drugSEprep(druglist = NULL, load_dir)
}
\arguments{
\item{druglist}{drug and its corresponding "drugbank ID" in \code{data
frame}}

\item{load_dir}{path to load or save modeling data files}
}
\value{
\code{drug_SE} a data frame contains drugs and their side effect
  terms. If druglist contains newly inputed drug, their drug-sideeffect
  dataframe will be merged with provided data
}
\description{
The function \code{drugSEprep} is used to prepare input file for
  function \code{SE.tanimoto}
}
\details{
\code{drugSEprep} provides drug-side effect file download from
  \code{SIDER4} database, where newly inputed drugs can be matched to their
  known side effect terms through "drugbank ID".
}
\examples{

\dontrun{
load_dir = "G:/lab/DCcomboNet/Rpackage/input_data/"
druglist = NULL
# or druglist = read.table("druglist_example.csv", sep = ",", header = TRUE,
# stringsAsFactors = FALSE)
drugSEprep(druglist = druglist, load_dir = load_dir)
}

}
