% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Potential_mechanism_analysis.R
\name{DT.Dis}
\alias{DT.Dis}
\title{Drug similarity based on Drug target distance}
\usage{
DT.Dis(drug_target, PPI)
}
\arguments{
\item{drug_target}{drug-target gene interaction file in \code{data.frame}
format (otherwise \code{as.data.frame} can be used to transform the format
of file). Target gene names should convert to official gene name (i.e. HGNC
gene symbol)}

\item{PPI}{protein-protein interaction network. User can use either the
default network from inBioMap PPI network or offer your own PPT network}
}
\value{
Drug average target gene distance matrix
}
\description{
The function \code{DT.Dis} is to calculate the drug-drug
  similarity based on drug target distance in PPI network.
}
