% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugNetFeatures.R
\name{FP.tanimoro}
\alias{FP.tanimoro}
\title{Tanimoto Coefficient between Drug Chemical Structure fingerprints}
\usage{
FP.tanimoro(fingerprint)
}
\arguments{
\item{fingerprint}{Drug fingerprint matrix (binary 0-1 matrix) generated via
PEDAL, rows denote drugs and columns denote particular substructures in the
molecule.}
}
\value{
tanimoto_score.matrix
}
\description{
The function \code{FP.tanimoro} is used for computing Tanimoto
  Coefficient between drug chemical structure fingerprints. The Chemical
  fingerprints can be generated from multiple softwares. Here we use PEDAL
  generated fingerprint file.
}
\examples{
\dontrun{
load_dir = "G:/lab/DCcomboNet/Rpackage/input_data/"
fingerprint = read.table(paste0(load_dir,
"/data/fingerprints/fingerprints.csv"), sep = ",", header = TRUE,
stringsAsFactors = FALSE)
FP.tanimoro(fingerprint = fingerprint)
}

}
