% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneNet.R
\name{geneNet.Adj}
\alias{geneNet.Adj}
\title{Generate the adjacency matrix from gene-gene interaction networks}
\usage{
geneNet.Adj(GeneNetwork)
}
\arguments{
\item{GeneNetwork}{\code{igraph} network file generated from
\code{\link{geneNet}}}
}
\value{
\code{adjMatrix} gene-gene association network adjacency matrix
  (sparse matrix format)
}
\description{
The function \code{geneNet.Adj} is used to generate adjacency
  matrix from gene-gene interaction network built from \code{\link{geneNet}}
}
\examples{
# require(igraph)
g <- data.frame(V1 = c("g1", "g2", "g3", "g4"), V2 = c("g5", "g2", "g6",
"g1"))
g.net <- igraph::graph.data.frame(g)
geneNet.Adj(g.net)
}
\seealso{
\code{\link{geneNet}}
}
