% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{genepathway}
\alias{genepathway}
\title{KEGG pathway and genes within pathways}
\description{
Table of KEGG pathway and genes within pathways. The table contains three
columns, first column denotes the KEGG pathway ID, second column denotes the
names of these pathways (all letter in uppercase), third column denotes genes
within these pathways.
}
