% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOOCV_new.R
\name{L1.LOCOCV}
\alias{L1.LOCOCV}
\title{L1.LOCOCV}
\usage{
L1.LOCOCV(
  load_dir,
  resultdir,
  model = "L1",
  drugcandidate = NULL,
  drugnetWeight = c("TRUE", "T"),
  featuretype = "integrated_score",
  drugtarget,
  druggene = NULL,
  cancergene = NULL,
  dtweight = 2,
  dgweight = 1,
  dgpweight = 1,
  gpweight = 1,
  x = 0.5,
  y = 0.5,
  z = 0,
  A = 0.5,
  B = 0.5,
  eta = 1,
  r = 0.7
)
}
\arguments{
\item{load_dir}{path to load or save modelin data files}

\item{resultdir}{path to save result files}

\item{model}{choose model type, two models here to choose: "L1", "L2"}

\item{drugcandidate}{drug and its corresponding "drugbank ID" in \code{data
frame}}

\item{drugnetWeight}{\code{TRUE/FALSE}, choose if the edge of drug network
should be weighted by feature values. The abbreviationS "\code{T/F}" are
also acceptable}

\item{featuretype}{if parameter "\code{weighted = TRUE/T}", choose the
feature type to weight the edge of drug-drug network. The options are:
"ATCsim","STsim","SEsim","integrated_score","integrated_score2"}

\item{drugtarget}{drug-target gene interaction file in \code{data.frame}
format (otherwise \code{as.data.frame} can be used to transform the format
of file). Target gene names should convert to official gene name (i.e. HGNC
gene symbol)}

\item{druggene}{drug-gene interaction file (Optional). drug and genes that
either induced by drug or may influence drug effect can be provided for
construct gene network for level one model. Genes can be collected from
publications or from IPA tool. Gene name should be converted to official
gene names (i.e. HGNC gene symbols)}

\item{cancergene}{cancer_gene a list of cancer related genes. For level one
model here, \code{cancer_gene} list is provided within the function, it can
be \code{NULL}.}

\item{dtweight}{the weight of drug-target gene edges}

\item{dgweight}{the weight of drug-related gene edge}

\item{dgpweight}{numeric, weight of drug-pathway associations through drug
related genes}

\item{gpweight}{the weight of gene-pathway edges}

\item{x}{numeric, the probability that the inter-network crossing event
happens between drug and gene nodes}

\item{y}{numeric, the probability that the inter-network crossing event
happens between drug and pathway nodes}

\item{z}{numeric, the probability that the inter-network crossing event
happens between gene and pathway nodes}

\item{A}{numeric, the probability of jumping within drugnet, when the
inter-network crossing events between drug and gene nodes and between drug
and pathway nodes are neither existing}

\item{B}{numeric, the probability of jumping within genenet, when the
inter-network crossing events between drug and gene nodes and between gene
and pathway nodes are neither existing}

\item{eta}{numeric parameter to controls the probability of restarting in the
corresponding network}

\item{r}{numeric, global restart parameter}

\item{CDK_FP}{Drug fingerprint matrix generated from PEDAL}

\item{pubchemFP}{Drug Pubchem fingerprint matrix generated from PEDAL}

\item{MACCS_FP}{Drug MACCS fingerprint matrix generated from PEDAL}
}
\value{
Ranked result tables of global proximity between drug seed and other
  drug nodes computed from drug-gene/pathway interaction network. File will
  be saved under the path \code{resultdir} provided by user.
}
\description{
the function "L1.LOCOCV" is to predict possible drug to pair
  with the drug candidate without expression data
}
