% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_extraction_visualization.R
\name{network_extract}
\alias{network_extract}
\title{Extract predicted drug combination subnetwork}
\usage{
network_extract(
  drugseed,
  drugcandidate,
  drugtarget = NULL,
  generank_filter = 0.01,
  pathwayrank_filter = 0.1,
  model = c("L1", "L2"),
  cellline = NULL,
  load_dir,
  resultdir
)
}
\arguments{
\item{drugseed}{drug that user is interested in perdicting combinable drug(s)}

\item{drugcandidate}{predicted combinable drug for drugseed}

\item{drugtarget}{drug-target gene interaction file corresponding to
\code{drugseed} and \code{drugcandidate} in \code{data.frame} format
(otherwise \code{as.data.frame} can be used to transform the format of
file). Target gene names should convert to official gene name (i.e. HGNC
gene symbol).}

\item{generank_filter}{numeric, to select genes according to the rank,
default value is 0.01 meaning only genes ranked on top 1% will be kept in
subnetwork}

\item{pathwayrank_filter}{umeric, to select pathways according to the rank,
default value is 0.1 meaning only pathways ranked on top 10% will be kept in
subnetwork}

\item{model}{choose model type, two models here to choose: "L1", "L2"}

\item{cellline}{cancer cell lines name.}

\item{load_dir}{path to load or save modelin data files}

\item{resultdir}{path to save result files}
}
\value{
Predicted drug combination subnetwork table extracted from
  drug-gene/pathway network constructed based on user requirements
}
\description{
Function \code{network_extract} is to extract subnetwork of
  predicted drug combinations from drug-gene/pathway network to help infer
  possible mechanism of drug combinations and visualization. Note that this
  function only works after user did prediction via \code{DComboNet()} and
  drug/gene/pathway rank tables have been generated and saved in fetchable
  path provided in \code{resultdir}.
}
\examples{
drugseed = "Sorafenib"
drugcandidate = "Vorinostat"
drugtarget = data.frame(Drug = c(rep(drugseed,10),rep(drugcandidate,5)),target
= c("BRAF", "FGFR1", "FLT1", "FLT3", "FLT4", "KDR", "KIT", "PDGFRB", "RAF1",
"RET", "HDAC1", "HDAC2", "HDAC3", "HDAC6", "HDAC8"))
load_dir = "G:/lab/DCcomboNet/Rpackage/input_data/"
resultdir = "G:/lab/DCcomboNet/Rpackage/tryout_result/"
\dontrun{network_extract(drugseed = drugseed, drugcandidate = drugcandidate,
drugtarget = drugtarget, generank_filter = 0.01, pathwayrank_filter = 0.1,
model = "L1", load_dir = load_dir, resultdir = resultdir)}
}
