% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_extraction_visualization.R
\name{network_visualization}
\alias{network_visualization}
\title{Predicted drug combination subnetwork visualization}
\usage{
network_visualization(
  drugseed,
  drugcandidate,
  drugtarget = NULL,
  model = c("L1", "L2"),
  cellline = NULL,
  load_dir,
  resultdir
)
}
\arguments{
\item{drugseed}{drug that user is interested in perdicting combinable drug(s)}

\item{drugcandidate}{predicted combinable drug for drugseed}

\item{drugtarget}{drug-target gene interaction file in \code{data.frame}
format (otherwise \code{as.data.frame} can be used to transform the format
of file). Target gene names should convert to official gene name (i.e. HGNC
gene symbol)}

\item{model}{choose model type, two models here to choose: "L1", "L2"}

\item{cellline}{cancer cell lines name.}

\item{load_dir}{path to load or save modelin data files}

\item{resultdir}{path to save result files}
}
\value{
Interactive subnetwork graph of predicted drug combination. An
  \code{.graphml} network file will be saved in provided path and can be
  easily import into \code{Cytoscape} for further analysis.
}
\description{
Function \code{network_visualization} is to visualize subnetwork
  of predicted drug combinations from drug-gene/pathway network to help infer
  possible mechanism of drug combinations. Note that subnetwork table should
  be prepared via function \code{network_extract}.
}
\examples{
drugseed = "Sorafenib"
drugcandidate = "Vorinostat"
drugtarget = data.frame(Drug = c(rep(drugseed,10),rep(drugcandidate,5)),target
= c("BRAF", "FGFR1", "FLT1", "FLT3", "FLT4", "KDR", "KIT", "PDGFRB", "RAF1",
"RET", "HDAC1", "HDAC2", "HDAC3", "HDAC6", "HDAC8"))

load_dir = "G:/lab/DCcomboNet/Rpackage/input_data/"
resultdir = "G:/lab/DCcomboNet/Rpackage/tryout_result/"

\dontrun{network_visualization(drugseed = drugseed, drugcandidate =
drugcandidate, drugtarget = drugtarget, model = "L2", load_dir = load_dir,
resultdir = resultdir)}

}
