% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayNet.R
\name{pathwayNet.Adj}
\alias{pathwayNet.Adj}
\title{Generate the adjacency matrix from pathway-pathway association network}
\usage{
pathwayNet.Adj(PathwayNetwork)
}
\arguments{
\item{PathwayNetwork}{\code{data.frame} of pathway-pathway associations}
}
\value{
\code{pathwayadj} pathway-pathway association adjacency matrix
  (sparse matrix)
}
\description{
The function \code{pathwayNet.Adj} is to generate adjacency
  matrix from pathway-pathway interaction network
}
\examples{
# require(igraph)
p <- data.frame(V1 = c("p1", "p2", "p3", "p4"), V2 = c("p5", "p2", "p6",
"p1"))

pathwayNet.Adj(p)

}
