% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{PPI_table}
\alias{PPI_table}
\title{Protein-protein interaction network}
\description{
Protein-protein interaction network table \code{PPI_table}, extracted from inBio map. This
table contains protein-protein interactions and the confidence scores
supporting the existence of edges.
}
