% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Results_rank.R
\name{rank_drugs}
\alias{rank_drugs}
\title{Drug result rank}
\usage{
rank_drugs(Num.Gene, Num.Drug, RWR.result, Drug_seeds)
}
\arguments{
\item{Num.Gene}{Number of nodes in gene-gene association network}

\item{Num.Drug}{Number of nodes in drug-drug association network}

\item{RWR.result}{data.frame, global proximity between drug seed and other
drug in drug-gene/pathway interaction network computed via \code{rwr}
function}

\item{Drug_seeds}{drug that take as seed in \code{seedscore} function}
}
\value{
sorted global proximity between drug seed and other drug in
  drug-gene/pathway interaction network
}
\description{
The function \code{rank_drugs} is to rank drugs according to rwr
  result
}
