% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Results_rank.R
\name{rank_genes}
\alias{rank_genes}
\title{Gene result rank}
\usage{
rank_genes(Num.Drug, Num.Gene, RWR.result, Gene_seeds = NULL)
}
\arguments{
\item{Num.Drug}{Number of nodes in drug-drug association network}

\item{Num.Gene}{Number of nodes in gene-gene association network}

\item{RWR.result}{data.frame, global proximity between drug seed and other
drug in drug-gene/pathway interaction network computed via \code{rwr}
function}

\item{Gene_seeds}{cancer related gene seeds (optional)}
}
\value{
global proximity between drug seed and other
  genes in drug-gene/pathway interaction network
}
\description{
The function \code{rank_genes} is to rank genes according to rwr result
}
