% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Results_rank.R
\name{rank_pathways}
\alias{rank_pathways}
\title{Pathway result rank}
\usage{
rank_pathways(Num.Gene, Num.Drug, RWR.result)
}
\arguments{
\item{Num.Gene}{Number of nodes in gene-gene association network}

\item{Num.Drug}{Number of nodes in drug-drug association network}

\item{RWR.result}{data.frame, global proximity between drug seed and other
drug in drug-gene/pathway interaction network computed via \code{rwr}
function}
}
\value{
global proximity between drug seed and other pathways in
  drug-gene/pathway association network
}
\description{
The function \code{rank_pathways} is to rank pathways according
  to rwr result
}
