% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prediction.R
\name{Ranking}
\alias{Ranking}
\title{Ranking system of DComboNet}
\usage{
Ranking(
  drugcandidate,
  model,
  resultdir,
  filename = NULL,
  outputdir = NULL,
  methods = c("sum", "two_threshold"),
  threshold_syn = 0.1,
  threshold_ant = 0.5
)
}
\arguments{
\item{drugcandidate}{drug and its corresponding "drugbank ID" in
\code{data.frame}}

\item{model}{choose model type, two models here to choose: "L1", "L2"}

\item{resultdir}{path to where result files saved}

\item{methods}{choose method used to set the final rank. Two methods are
are available to choose: "sum" and "two_threshold". Default method is
"sum".}

\item{threshold_syn}{if \code{methods = "two_threshold"}, the threshold used
to determine the combinable pairs.}

\item{threshold_ant}{if \code{methods = "two_threshold"}, the threshold used
to determine the uncombinable pairs.}
}
\value{
Ranked result tables. Each drug seed will create three files that
  contains global proximity between drug seed and other drug/gene/pathway
  nodes computed from drug-gene/pathway interaction network. File will be
  saved under the path \code{resultdir} provided by user.
}
\description{
the function \code{DComboNet} is to predict possible drug to
  pair with the drug candidate without expression data
}
\examples{
\dontrun{
result_dir = 'G:/lab/Projects/p1_DComboNet/DComboNet_improve/OCI_LY3_tryout_results/result_overlap/'
outputdir = 'G:/lab/Projects/p1_DComboNet/DComboNet_improve/OCI_LY3_tryout_results/'
Ranking(
 drugcandidate = druglist,
 model = "L2",
 resultdir = result_dir,
 filename = 'test',
 outputdir = outputdir,
 methods = "sum",
 threshold_syn = 0.1,
 threshold_ant = 0.5
)}

}
