% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugNetFeatures.R
\name{SE.sim}
\alias{SE.sim}
\title{Tanimoto Coefficient between Drug side effect}
\usage{
SE.sim(druglist = NULL, load_dir)
}
\arguments{
\item{druglist}{a dataframe of drug and corresponding "drugbank ID"}

\item{load_dir}{path to load and save modeling data files}
}
\value{
\code{SE.tanimoto.matrix} Drug similarity matrix based on drug side
  effects Tanimoto Coefficient.
}
\description{
The function \code{SE.sim} is used to generate drug side effect
  based similarity matrix
}
\details{
drug-side effect dataframe will be converted into 0-1 binary matrix
  where rows denote drugs columns denote side effect terms, 1 represent drug
  has corresponding side effect term extracted from \code{SIDER4} database
  while 0 represent does not has.  Tanimoto coefficient was considered to
  compute side effect-based similarity.
}
\examples{

\dontrun{
load_dir = "G:/lab/DCcomboNet/Rpackage/input_data/"
druglist = NULL
# or druglist = read.table("druglist_example.csv", sep = ",", header = TRUE,
# stringsAsFactors = FALSE)
SE.sim(druglist = druglist, load_dir = load_dir)
}

}
