% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeedsPreparation.R
\name{seedscore}
\alias{seedscore}
\title{Generate drugseed-score file}
\usage{
seedscore(seeds, eta = 1)
}
\arguments{
\item{seeds}{strings, randow walk with restart process will start from the
provided drug seed which should be included in drug network}

\item{eta}{numeric parameter to controls the probability of restarting in the
corresponding network}
}
\value{
\code{seeds_score} drugseed-score \code{data.frame}
}
\description{
The function \code{seedscore} is to generate the seed-score file
}
\examples{
seeds = "Sorafenib"
seedscore(seeds = seeds)
}
