% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugNetFeatures.R
\name{structure.sim}
\alias{structure.sim}
\title{Drug Chemical Fingerprint Similarity Function}
\usage{
structure.sim(CDK_FP = NULL, pubchemFP = NULL, MACCS_FP = NULL, load_dir)
}
\arguments{
\item{CDK_FP}{Drug fingerprint matrix generated from PEDAL}

\item{pubchemFP}{Drug Pubchem fingerprint matrix generated from PEDAL}

\item{MACCS_FP}{Drug MACCS fingerprint matrix generated from PEDAL}

\item{load_dir}{path to load and save model data file(s)}
}
\value{
\code{structure_sim} drug-drug chemical fingerprint-based similarity
  matrix, rows and columns represent drug names
}
\description{
The function \code{structure.sim} is for computing drug chemical
  structure fingerprint based similarity.
}
\details{
Three kinds of chemical fingerprint based similarity will be computed
 and integrated into one score. Tanimoto index is used to calculate the
 similarity matrix. The Chemical fingerprints can be generated from multiple
 softwares, here we use PEDAL to generate three kinds fingerprint matrix,
 including CDK fingerprint, pubchem fingerprint and MACCS fingerprint. The
 average score of these three fingerprint based similarity is computed as
 final drug-drug chemical fingerprint-based similarity. Four similarity
 matrices will be calculated and saved to path provided by user
 (\code{load_dir}) under \emph{"/data/fingerprints/"} folder.
}
\examples{
\dontrun{
load_dir = "G:/lab/DCcomboNet/Rpackage/input_data/"
structure.sim(load_dir = load_dir)
}


}
