% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Potential_mechanism_analysis.R
\name{Targets.Dis}
\alias{Targets.Dis}
\title{Drug similarity based on Drug target distancce}
\usage{
Targets.Dis(targets, PPI, load_dir)
}
\arguments{
\item{targets}{target genes list for drug candidates, target gene names
should convert to official gene name (i.e. HGNC gene symbol)}

\item{PPI}{protein-protein interaction network. User can use either the
default network from inBioMap PPI network or offer your own PPT network}

\item{load_dir}{path to load or save modelling data files}
}
\value{
Distance matrix between target genes in protein-protein interaction
  network.
}
\description{
The function \code{Targets.Dis} is to generate the distance
  matrix between drug targets
}
