% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugGeneHeteroNet.R
\name{TransitionMatrix}
\alias{TransitionMatrix}
\title{Generate transition matrix from drug-gene/pathway interaction networks}
\usage{
TransitionMatrix(
  drugAdj,
  geneAdj,
  pathwayAdj,
  druggeneAdj,
  drugpathwayAdj,
  genepathwayAdj,
  x = 0.5,
  y = 0.5,
  z = 0,
  A = 0.5,
  B = 0.5
)
}
\arguments{
\item{drugAdj}{drug adjacency matrix (sparse matrix) generated from drug-drug
association network via \code{AdjMatrix_drugs}}

\item{geneAdj}{adjacency matrix of gene-gene association network (sparse
matrix) via \code{geneNet.Adj}}

\item{pathwayAdj}{adjacency matrix of pathway-pathway association network
(sparse matrix) via \code{pathwayNet.Adj}}

\item{druggeneAdj}{adjacency matrix of drug-gene association network (sparse
matrix) via \code{drugGeneNet.L1} or \code{drugGeneNet.L2} depends on which
model is called}

\item{drugpathwayAdj}{adjacency matrix of drug-pathway association network
(sparse matrix) via \code{DrugPathwayMatrix.L1} or
\code{DrugPathwayMatrix.L2} depends on which model is called}

\item{genepathwayAdj}{adjacency matrix of gene-pathway association network
(sparse matrix) via \code{genepathwayAdj}}

\item{x}{numeric, the probability that the inter-network crossing event
happens between drug and gene nodes}

\item{y}{numeric, the probability that the inter-network crossing event
happens between drug and pathway nodes}

\item{z}{numeric, the probability that the inter-network crossing event
happens between gene and pathway nodes}

\item{A}{numeric, the probability of jumping within drugnet, when the
inter-network crossing events between drug and gene nodes and between drug
and pathway nodes are neither existing}

\item{B}{numeric, the probability of jumping within genenet, when the
inter-network crossing events between drug and gene nodes and between gene
and pathway nodes are neither existing}
}
\value{
Transition matrix generated from drug-gene/pathway interaction
  networks
}
\description{
The function \code{TransitionMatrix} is to generate the
  adjacency matrix from drug-drug association network, drug-gene association
  network, gene-gene association network, drug-pathway association network,
  pathway-pathway association network and gene-pathway association network.
}
\examples{
\dontrun{
# This function replies on adjacency matrices from different subnetworks
  drugAdj = AdjMatrix_drugs(...)
  geneAdj = geneNet.Adj(...)
  pathwayAdj = pathwayNet.Adj(...)
  # For level one model
  druggeneAdj = drugGeneNet.L1(...)
  drugpathwayAdj = DrugPathwayMatrix.L1(...)
  # For level two model
  druggeneAdj = drugGeneNet.L2(...)
  drugpathwayAdj = DrugPathwayMatrix.L2(...)
  genepathwayAdj = genepathwayAdj(...)
  x = 0.5
  y = 0.5
  z = 0
  A = 0.5
  B = 0.5
  TransitionMatrix(drugAdj = drugAdj, geneAdj = geneAdj, pathwayAdj =
  pathwayAdj, druggeneAdj = druggeneAdj, drugpathwayAdj = drugpathwayAdj,
  genepathwayAdj = genepathwayAdj, x = 0.5, y = 0.5, z = 0, A = 0.5, B = 0.5)
}
}
\seealso{
Functions to generate adjacency matrix for different subnetworks:
  \code{\link{AdjMatrix_drugs}} for drug-drug association network,
  \code{\link{geneNet.Adj}} for gene-gene association network,
  \code{\link{pathwayNet.Adj}} for pathway-pathway association network,
  \code{\link{drugGeneNet.L1}} or \code{\link{drugGeneNet.L2}} for drug-gene
  association network, \code{\link{DrugPathwayMatrix.L1}} or
  \code{\link{DrugPathwayMatrix.L2}} for drug-pathway association network,
  \code{\link{genepathwayAdj}} for gene-pathway association network.
}
