% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Potential_mechanism_analysis.R
\name{WWI.sim}
\alias{WWI.sim}
\title{Drug similarity based on pathway similarity}
\usage{
WWI.sim(drugpair, drugtarget, targetPW, PPI)
}
\arguments{
\item{drugpair}{drug pairs generated from \code{druglist} you provided (or
the subset)}

\item{drugtarget}{drug-target gene interaction file in \code{data.frame}
format (otherwise \code{as.data.frame} can be used to transform the format
of file). Target gene names should convert to official gene name (i.e. HGNC
gene symbol)}

\item{targetPW}{gene and their related pathways table}

\item{PPI}{protein-protein interaction network. You can either use the
default network from INBIO or offer your own PPT network.}
}
\value{
drug.matrix Drug-drug similarity matrix based on the pathway-pathway
  interaction
}
\description{
The function \code{WWI.sim} is to generate the distance matrix
  between drug targets
}
