% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{WWI_all}
\alias{WWI_all}
\title{KEGG pathway-pathway interaction network}
\description{
Table of KEGG pathway-pathway interaction network downloaded from publication
(PMID: 26687590). The first two columns denote the interaction between two
pathway nodes; the third column denotes the interaction types.
}
